# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: GCRF Library Management System API
  version: 1.0.0
  description: |
    RESTful API documentation for GCRF (国创睿峰) Intelligent Library Management System.

    This system provides comprehensive library management capabilities including:
    - Authentication and authorization
    - Book catalog management
    - Reader management
    - Circulation (borrow, return, renew)
    - System administration

    **Architecture**: Microservices with Spring Cloud Gateway

    **Base URL**: All requests go through the API Gateway at `http://localhost:8080`

    **Authentication**: JWT Bearer token required for most endpoints (see Authentication section)

  contact:
    name: GCRF Library Team
    email: support@gcrf-library.com
  license:
    name: Proprietary
    url: https://gcrf-library.com/license
servers:
  - url: http://localhost:8080
    description: Development server (Local)
  - url: https://api.gcrf-library.com
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/roles:
    get:
      tags: [Roles]
      summary: List roles with pagination
      description: Query role list with filters (Admin only)
      parameters:
        - name: pageNum
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
        - name: roleCode
          in: query
          description: Role code filter
          schema:
            type: string
        - name: roleName
          in: query
          description: Role name filter
          schema:
            type: string
        - name: status
          in: query
          description: Role status
          schema:
            type: string
            enum: [ACTIVE, INACTIVE]
      responses:
        '200':
          description: Role list retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PagedRoleResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags: [Roles]
      summary: Create role
      description: Create a new role (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleCreateRequest'
      responses:
        '200':
          description: Role created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RoleDetailVO'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/v1/roles/{id}:
    get:
      tags: [Roles]
      summary: Get role details
      description: Retrieve role details with permissions
      parameters:
        - name: id
          in: path
          required: true
          description: Role ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Role details retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RoleDetailVO'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags: [Roles]
      summary: Update role
      description: Update role information (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          description: Role ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleUpdateRequest'
      responses:
        '200':
          description: Role updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RoleDetailVO'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags: [Roles]
      summary: Delete role
      description: Soft delete role (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          description: Role ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Role deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/roles/{id}/permissions:
    get:
      tags: [Roles]
      summary: Get role permissions
      description: Retrieve list of permissions assigned to role
      parameters:
        - name: id
          in: path
          required: true
          description: Role ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Permissions retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/PermissionVO'

    post:
      tags: [Roles]
      summary: Assign permissions to role
      description: Assign multiple permissions to a role (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          description: Role ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - permissionIds
              properties:
                permissionIds:
                  type: array
                  items:
                    type: integer
                    format: int64
                  description: List of permission IDs to assign
                  example: [1, 2, 3, 5, 8]
      responses:
        '200':
          description: Permissions assigned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ========================================
  # Health Check APIs
  # ========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication. Include the token in the Authorization header:
        ```
        Authorization: Bearer <your-jwt-token>
        ```
...
  schemas:
...
    RoleCreateRequest:
      type: object
      required:
        - roleCode
        - roleName
      properties:
        roleCode:
          type: string
          maxLength: 50
          example: "LIBRARIAN"
        roleName:
          type: string
          maxLength: 100
          example: "图书管理员"
        description:
          type: string
          maxLength: 500
          example: "负责图书入库、借还等日常管理"
        sortOrder:
          type: integer
          example: 10
    RoleUpdateRequest:
      type: object
      properties:
        roleName:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 500
        sortOrder:
          type: integer
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
    RoleDetailVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        roleCode:
          type: string
          example: "LIBRARIAN"
        roleName:
          type: string
          example: "图书管理员"
        description:
          type: string
          example: "负责图书入库、借还等日常管理"
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
          example: "ACTIVE"
        sortOrder:
          type: integer
          example: 10
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PermissionVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        permissionCode:
          type: string
          example: "book:create"
        permissionName:
          type: string
          example: "创建图书"
        permissionType:
          type: string
          enum: [MENU, BUTTON, API]
          example: "API"
        resourcePath:
          type: string
          example: "/api/v1/books"
        description:
          type: string

    # ========================================
    # Response Wrappers
    # ========================================
    SuccessResponse:
      type: object
      properties:
        code:
          type: integer
          description: Response code (200 for success)
          example: 200
        message:
          type: string
          description: Response message
          example: "操作成功"
        data:
          description: Response data (varies by endpoint)
        timestamp:
          type: integer
          format: int64
          description: Unix timestamp in milliseconds
          example: 1698765432000
        traceId:
          type: string
          description: Request trace ID for debugging
          example: "trace-abc123xyz"
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          description: Error code
          example: 400
        message:
          type: string
          description: Error message
          example: "参数验证失败"
        data:
          nullable: true
        timestamp:
          type: integer
          format: int64
          example: 1698765432000
        traceId:
          type: string
          example: "trace-abc123xyz"
    PageResult:
      type: object
      properties:
        records:
          type: array
          items: {}
          description: List of records
        total:
          type: integer
          format: int64
          description: Total number of records
          example: 100
        pageNum:
          type: integer
          description: Current page number (1-based)
          example: 1
        pageSize:
          type: integer
          description: Page size
          example: 10
        pages:
          type: integer
          description: Total number of pages
          example: 10
...
    PagedRoleResponse:
      allOf:
        - $ref: '#/components/schemas/PageResult'
        - type: object
          properties:
            records:
              type: array
              items:
                $ref: '#/components/schemas/RoleDetailVO'
...
  responses:
    UnauthorizedError:
      description: Authentication required or token invalid/expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missingToken:
              summary: Missing token
              value:
                code: 401
                message: "未提供认证令牌"
                data: null
                timestamp: 1698765432000
            invalidToken:
              summary: Invalid token
              value:
                code: 401
                message: "令牌无效或已过期"
                data: null
                timestamp: 1698765432000
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              value:
                code: 404
                message: "请求的资源不存在"
                data: null
                timestamp: 1698765432000
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation:
              value:
                code: 400
                message: "参数验证失败：用户名不能为空"
                data: null
                timestamp: 1698765432000
...